% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnrec_colors.R
\docType{data}
\name{dnrec_colors}
\alias{dnrec_colors}
\title{List of colors and names from the DNREC Brand and Editorial Guide, July 2021}
\format{
An object of class \code{character} of length 19.
}
\usage{
dnrec_colors
}
\description{
List of colors and names from the DNREC Brand and Editorial Guide, July 2021
}
\references{
DNREC Brand and Editorial Guide, July 2021.
}
\author{
Brittney Flaten \email{brittney.flaten@delaware.gov}
}
\keyword{data}
