% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnrec_cols.R
\name{dnrec_cols}
\alias{dnrec_cols}
\title{Obtain hex code from color name}
\usage{
dnrec_cols(color, ...)
}
\arguments{
\item{color}{Name of color from 2021 Guide. No capitals.}

\item{...}{Additional Arguments}
}
\value{
hex code or list of codes
}
\description{
Obtain hex code from color name
}
\examples{
dnrec_cols("bayshore green")
dnrec_cols("golden yellow", "moss green")
}
