% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnrec_pal.R
\name{dnrec_pal}
\alias{dnrec_pal}
\title{Color palettes based on DNREC Brand and Editorial Style Guide}
\usage{
dnrec_pal(palette = "brand", alpha = 1, ...)
}
\arguments{
\item{palette}{Character name of palette from the list of palettes.}

\item{alpha}{Numeric from 0 to 1 representing transparency.}

\item{...}{Additional arguments}
}
\value{
Palettes to be used in scale functions.
}
\description{
Color palettes based on DNREC Brand and Editorial Style Guide
}
