% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnrec_pals.R
\docType{data}
\name{dnrec_pals}
\alias{dnrec_pals}
\title{List of palettes from the DNREC Brand and Editorial Guide, July 2021}
\format{
An object of class \code{list} of length 5.
}
\usage{
dnrec_pals
}
\description{
List of palettes from the DNREC Brand and Editorial Guide, July 2021
}
\references{
DNREC Brand and Editorial Guide, July 2021.
}
\author{
Brittney Flaten \email{brittney.flaten@delaware.gov}
}
\keyword{data}
