% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnrec_scale_color.R
\name{dnrec_scale_color}
\alias{dnrec_scale_color}
\title{DNREC Brand Colors color scale}
\usage{
dnrec_scale_color(palette = "brand", alpha = 1, ...)
}
\arguments{
\item{palette}{Palette name. Brand, secondary, tertiary, extended or mix.}

\item{alpha}{Numeric transparency value from 0 to 1.}

\item{...}{Additional arguments}
}
\value{

}
\description{
DNREC Brand Colors color scale
}
\examples{
ggplot(iris, aes(Sepal.Length, Sepal.Width,
color=Species))+
geom_point(size=3)+dnrec_scale_color("tertiary")
}
