% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnrec_scale_fill.R
\name{dnrec_scale_fill}
\alias{dnrec_scale_fill}
\title{DNREC Brand Colors fill scale}
\usage{
dnrec_scale_fill(palette = "brand", alpha = 1, ...)
}
\arguments{
\item{palette}{Palette name. Brand, secondary, tertiary, extended or mix.}

\item{alpha}{Numeric transparency value from 0 to 1.}

\item{...}{Additional arguments}
}
\value{

}
\description{
DNREC Brand Colors fill scale
}
\examples{
ggplot(iris, aes(Species, fill=Species))+
geom_histogram(stat="count")+dnrec_scale_fill()
ggplot(iris, aes(Sepal.Length, fill=Species))+
 geom_histogram()+dnrec_scale_fill(palette="mix")

}
